backend/requirements.txt
autopep8==1.5.4
dnspython==2.1.0
email-validator==1.1.2
grpcio==1.26.0
grpcio-tools==1.26.0
idna==3.1
pg8000==1.16.6
protobuf==3.14.0
psycopg2==2.8.6
pycodestyle==2.6.0
scramp==1.2.0
six==1.15.0
SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.36.8
testing.common.database==2.0.3
toml==0.10.2
